native: move OracleRequest to state package
It is used by multiple modules outside native
and produces unneeded dependencies.
